+++ AutoPatcher.Engine/Repository/IRepositoryInternal.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace AutoPatcher.Engine.Repository
+{
+    internal interface IRepositoryInternal : IRepository
+    {
+        void AddBuildArtifactsRange(IEnumerable<BuildArtifact> buildArtifacts);
+
+        void ClearBuildArtifacts();
+    }
+}