+++ AutoPatcher.Engine/State.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
+using System.Linq;
 using System.Threading.Tasks;
 using AutoPatcher.Engine.Abstractions;
 using AutoPatcher.Engine.Properties;
@@ -15,6 +16,9 @@ namespace AutoPatcher.Engine
         private const string CurrentRemoteBinPathRootSettingKey = "AutoPatch_Engine_CurrentRemoteBinPathRoot";
         private const string EmptyRevisionSuffix = ".empty";
         private const string StockRevisionSuffix = ".stockrevision";
+
+        private readonly IDictionary<string, BuildArtifact> sourceItemPathToBuildArtifactLookupTable = new Dictionary<string, BuildArtifact>();
+
         private string currentRemoteBinRoot;
 
         public State(IErrorDialogs errorDialogs, IFileDialogs fileDialogs, ISettingsManager settingsManager)
@@ -48,7 +52,9 @@ namespace AutoPatcher.Engine
             }
         }
 
-        public IRepository Repository { get; private set; }
+        public IRepository Repository => this.MutableRepository;
+
+        private IRepositoryInternal MutableRepository { get; set; }
 
         private IErrorDialogs ErrorDialogs { get; }
 
@@ -60,7 +66,7 @@ namespace AutoPatcher.Engine
         {
             UnloadRepository();
 
-            this.Repository = RepositoryLoader.CreateEmptyFile(this.ErrorDialogs, filePath);
+            this.MutableRepository = RepositoryLoader.CreateEmptyFile(this.ErrorDialogs, filePath);
 
             await SaveRepositoryAsync();
         }
@@ -68,9 +74,10 @@ namespace AutoPatcher.Engine
         public async Task LoadRepositoryAsync(string filePath)
         {
             this.UnloadRepository();
-            this.Repository = await Task.Run(() => RepositoryLoader.CreateFromFile(this.ErrorDialogs, filePath));
+            this.MutableRepository = await Task.Run(() => RepositoryLoader.CreateFromFile(this.ErrorDialogs, filePath));
 
-            RefreshBuildArtifactStates(this.Repository.BuildArtifacts);
+            this.RefreshBuildArtifactStates(this.Repository.BuildArtifacts);
+            this.AddBuildArtifactsRangeToLookupTable(this.Repository.BuildArtifacts);
         }
 
         public async Task SaveRepositoryAsync()
@@ -87,7 +94,8 @@ namespace AutoPatcher.Engine
         {
             if (this.Repository != null)
             {
-                this.Repository = null;
+                this.MutableRepository = null;
+                this.sourceItemPathToBuildArtifactLookupTable.Clear();
             }
         }
 
@@ -95,7 +103,7 @@ namespace AutoPatcher.Engine
         {
             Verify.IsNotNull(buildArtifacts, nameof(buildArtifacts));
 
-            if (!this.IsInPatchableState())
+            if (!this.IsInActionableState())
             {
                 return;
             }
@@ -110,7 +118,7 @@ namespace AutoPatcher.Engine
         {
             Verify.IsNotNull(buildArtifacts, nameof(buildArtifacts));
 
-            if (!this.IsInPatchableState(true))
+            if (!this.IsInActionableState(true))
             {
                 return;
             }
@@ -121,28 +129,107 @@ namespace AutoPatcher.Engine
             }
         }
 
-        public void ClearBuildArtifacts() => this.Repository.BuildArtifacts.Clear();
+        public void ClearBuildArtifacts()
+        {
+            this.MutableRepository.ClearBuildArtifacts();
+            this.sourceItemPathToBuildArtifactLookupTable.Clear();
+        }
 
         public void AddBuildArtifactsRange(IEnumerable<BuildArtifact> buildArtifacts)
         {
             Verify.IsNotNull(buildArtifacts, nameof(buildArtifacts));
 
             this.RefreshBuildArtifactStates(buildArtifacts);
-
-            foreach (var buildArtifact in buildArtifacts)
-            {
-                this.Repository.BuildArtifacts.Add(buildArtifact);
-            }
+            this.MutableRepository.AddBuildArtifactsRange(buildArtifacts);
+            this.AddBuildArtifactsRangeToLookupTable(buildArtifacts);
         }
 
         public void RefreshBuildArtifactStates()
         {
+            // TODO: should this be async?
             if (this.Repository == null)
             {
                 throw new InvalidOperationException("No repository loaded");
             }
 
             this.RefreshBuildArtifactStates(this.Repository.BuildArtifacts);
+        }
+
+        public async Task<IEnumerable<BuildArtifact>> GetDirtyBuildArtifactsAsync() => await Task.Run(() => GetDirtyBuildArtifacts());
+
+        private IEnumerable<BuildArtifact> GetDirtyBuildArtifacts()
+        {
+            if (!this.IsInActionableState(isSourceOp: true))
+            {
+                yield break;
+            }
+
+            IEnumerable<string> dirtySourceItemPaths = null;
+
+            try
+            {
+                dirtySourceItemPaths = this.GetDirtySourceItemPaths();
+            }
+            catch (Exception ex)
+            {
+                this.ErrorDialogs.ErrorDialog(
+                    string.Format(
+                        Resources.StringFailureGeneral,
+                        ex.Message));
+
+                yield break;
+            }
+
+            foreach (var path in dirtySourceItemPaths)
+            {
+                BuildArtifact artifact;
+
+                if (this.sourceItemPathToBuildArtifactLookupTable.TryGetValue(path, out artifact))
+                {
+                    yield return artifact;
+                }
+                else
+                {
+                    // TODO: aggregate all into a single dialog.
+                    this.ErrorDialogs.WarningDialog(
+                        string.Format(
+                            Resources.StringUnmappedFile,
+                            path));
+                }
+            }
+        }
+
+        private IEnumerable<string> GetDirtySourceItemPaths()
+        {
+            Debug.Assert(!string.IsNullOrEmpty(this.Repository.SourceItemsRoot));
+
+            var sourceItemPaths = new List<string>();
+
+            try
+            {
+                using (var repo = new LibGit2Sharp.Repository(this.Repository.SourceItemsRoot))
+                {
+                    // TODO: get new files as dirty items.
+                    foreach (var change in repo.Diff.Compare<LibGit2Sharp.TreeChanges>())
+                    {
+                        sourceItemPaths.Add(change.Path);
+
+                        if (change.OldPath != change.Path)
+                        {
+                            sourceItemPaths.Add(change.OldPath);
+                        }
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                this.ErrorDialogs.ErrorDialog(
+                    string.Format(
+                        Resources.StringFailureGeneral,
+                        ex.Message));
+            }
+
+            return sourceItemPaths;
         }
 
         private void RefreshBuildArtifactStates(IEnumerable<BuildArtifact> buildArtifacts)
@@ -163,18 +250,30 @@ namespace AutoPatcher.Engine
             }
         }
 
-        private bool IsInPatchableState(bool isPatchRevert = false)
+        private void AddBuildArtifactsRangeToLookupTable(IEnumerable<BuildArtifact> buildArtifacts)
+        {
+            // TODO: analyze perf impact.
+            foreach (var buildArtifact in this.Repository.BuildArtifacts)
+            {
+                foreach (var sourceItem in buildArtifact.SourceItems)
+                {
+                    this.sourceItemPathToBuildArtifactLookupTable.Add(sourceItem.LocalPath, buildArtifact);
+                }
+            }
+        }
+
+        private bool IsInActionableState(bool isRevert = false, bool isSourceOp = false)
         {
             if (this.Repository == null)
             {
                 throw new InvalidOperationException("No repository loaded");
             }
 
-            if (!isPatchRevert && !Directory.Exists(this.Repository.LocalBinRoot))
+            if (!isRevert && !Directory.Exists(this.Repository.LocalBinRoot))
             {
                 this.ErrorDialogs.ErrorDialog(
                     string.Format(
-                        Resources.StringPatchFailureInvalidLocalBinRoot,
+                        Resources.StringFailureInvalidLocalBinRoot,
                         this.CurrentRemoteBinRoot));
                 return false;
             }
@@ -183,8 +282,17 @@ namespace AutoPatcher.Engine
             {
                 this.ErrorDialogs.ErrorDialog(
                     string.Format(
-                        Resources.StringPatchFailureInvalidRemoteBinRoot,
+                        Resources.StringFailureInvalidRemoteBinRoot,
                         this.CurrentRemoteBinRoot));
+                return false;
+            }
+
+            if (isSourceOp && !Directory.Exists(this.Repository.SourceItemsRoot))
+            {
+                this.ErrorDialogs.ErrorDialog(
+                    string.Format(
+                        Resources.StringFailureInvalidSourceItemsRoot,
+                        this.Repository.SourceItemsRoot));
                 return false;
             }