+++ AutoPatcher/Models/IMultiSelectable.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace AutoPatcher.Models
+{
+    internal interface IMultiSelectable
+    {
+        event EventHandler ModelChangedSelection;
+
+        IList<object> Selected { get; }
+
+        void RaiseModelChangedSelectionEvent();
+    }
+}