+++ AutoPatcher/Views/MainWindow.xaml.cs
@@ -1,5 +1,5 @@
-﻿using System.Windows;
-using AutoPatcher.Engine.Repository;
+﻿using System.Linq;
+using System.Windows;
 using AutoPatcher.Models;
 
 namespace AutoPatcher.Views
@@ -9,21 +9,51 @@ namespace AutoPatcher.Views
     /// </summary>
     public partial class MainWindow : Window
     {
+        private bool isSelectionUpdateInProgress;
+
         public MainWindow()
         {
             InitializeComponent();
         }
 
+        private void Window_Loaded(object sender, RoutedEventArgs e)
+        {
+            var multiSelectable = ((IMultiSelectable)this.DataContext);
+            multiSelectable.ModelChangedSelection += MultiSelectable_ModelChangedSelection;
+        }
+
+        private void MultiSelectable_ModelChangedSelection(object sender, System.EventArgs e)
+        {
+            // HACK: update view from model.
+
+            this.isSelectionUpdateInProgress = true;
+
+            var multiSelectable = ((IMultiSelectable)this.DataContext);
+            this.BuildArtifactsListBox.SelectedItems.Clear();
+
+            foreach (var selectedItem in multiSelectable.Selected)
+            {
+                this.BuildArtifactsListBox.SelectedItems.Add(selectedItem);
+            }
+
+            this.BuildArtifactsListBox.Focus();
+
+            this.isSelectionUpdateInProgress = false;
+        }
+
         private void BuildArtifactsListBox_SelectionChanged(object sender, System.Windows.Controls.SelectionChangedEventArgs e)
         {
-            // HACK: update model.
-            var selectedArtifacts = ((MainWindowModel)this.DataContext).SelectedBuildArtifacts;
-
-            selectedArtifacts.Clear();
-
-            foreach (var selection in this.BuildArtifactsListBox.SelectedItems)
+            if (!isSelectionUpdateInProgress)
             {
-                selectedArtifacts.Add((BuildArtifact)selection);
+                // HACK: update model from view.
+                var selectedArtifacts = ((IMultiSelectable)this.DataContext).Selected;
+
+                selectedArtifacts.Clear();
+
+                foreach (var selection in this.BuildArtifactsListBox.SelectedItems)
+                {
+                    selectedArtifacts.Add(selection);
+                }
             }
         }
     }